Moved notification checking logic out of the main Wuja application and into it's own class.
Implements the observer pattern so it can notify the main app (or whoever happens to be listening) when a notification needs to be displayed to the user. Not sure how much sense this makes right now, as somebody (i.e. the observer) has to tell the notifier when to do his checks, didn't want to get into spawning a seperate thread right now.

Fixed bug in the entry classes, one was returning a list of Event objects, the other datetime objects. Events are now always returned, and contain a reference to the calendar entry they're associated with.